chore: moved all ts files for examples to examples/ts
