+++ site/examples/ts/code-highlighting.tsx
@@ -32,8 +32,8 @@ import { withHistory } from 'slate-history'
 import isHotkey from 'is-hotkey'
 import { css } from '@emotion/css'
 import { CodeBlockElement } from './custom-types.d'
-import { normalizeTokens } from '../utils/normalize-tokens'
-import { Button, Icon, Toolbar } from '../components'
+import { normalizeTokens } from './utils/normalize-tokens'
+import { Button, Icon, Toolbar } from './components'
 
 const ParagraphType = 'paragraph'
 const CodeBlockType = 'code-block'
