+++ site/examples/ts/components/index.tsx
