+++ site/examples/ts/custom-types.d.ts
