+++ site/examples/ts/forced-layout.tsx
