+++ site/examples/ts/markdown-shortcuts.tsx
