+++ site/examples/ts/plaintext.tsx
