+++ site/examples/ts/read-only.tsx
